Provide seed control for datasets when shuffling
Setting the seed on the server is not enough to control the randomness of the server in
an experiment. Indeed, there might be multiple calls to get other challenges, so we need
to rely on RandomStates in functions instead of setting a global seed.
Integration tests should be added in the future to confirm the implementation provides
a deterministic server output.