Refactor model definition inside other modules
This will allow us to decouple the backend to the actual training platform.
A simple ModelStore returns an instantiated model provided the library name and model name.
Weight initialization and parameter logging has also been delegated to these modules.
All this is achieved by an EvaluationModel base class, which will still change in the future, but works for now.