Fix bug where seed would be different for same experiment with different evaluation type
This critical bug invalidates previous experiments, as the type of the evaluation type was taken
into account when generating a seed, where it is what we mean to control.
This fix also includes a new intermediate package, used to communicate between client and server:
both use an intermediary class that serves as a DTO for a cleaner interface than sending string.
This class is serializable by pyzmq as direct classes, but methods for transforming to native dict
are provided.

This commit also fixes the lru_cache policy for faster testing by splitting the data splitting/
shuffling into its own function and applying the cache on that, rather than on the callback
of the server when accepting a data request (different object were instantiated, thus the cache
did not contain an entry.